MODEXPW-470 Fixed generics
